+++ src/components/DatePickerCanary/__stories__/DatePicker.stories.tsx
@@ -21,14 +21,14 @@ import {
   textFieldPropView,
   textFieldPropViewDefault,
 } from '../../TextField/TextField';
-import { DatePicker } from '../DatePickerCanary';
 import {
+  DatePicker,
   datePickerPropDropdownForm,
   datePickerPropDropdownFormDefault,
   datePickerPropType,
   datePickerPropTypeDefault,
   DatePickerPropValue,
-} from '../helpers';
+} from '../DatePickerCanary';
 
 import mdx from './DatePicker.docs.mdx';
 
@@ -72,6 +72,8 @@ const additionalControls = () => {
   return [<Button label="Кнопка" />, <Button label="Кнопка" />];
 };
 
+export const isTypeRange = (type: string) => type.indexOf('radge') !== -1;
+
 export function Playground() {
   const {
     form,
@@ -105,8 +107,11 @@ export function Playground() {
   const icon = withIcon ? IconCalendar : undefined;
 
   useEffect(() => {
+    console.log(type);
     setValue(null);
   }, [type]);
+
+  console.log(value);
 
   return (
     <div style={{ maxWidth: 500 }}>
@@ -124,17 +129,13 @@ export function Playground() {
         disabled={disabled}
         size={size}
         onChange={({ value }) => setValue(value)}
-        rightSide={icon}
+        leftSide={icon}
         events={events}
         locale={getByMap(localeMap, locale)}
         dateTimeView={dateTimeView}
         dropdownForm={dropdownForm}
         minDate={new Date(minDate)}
         maxDate={new Date(maxDate)}
-        {...(type === 'date-range' && {
-          endFieldRightSide: icon,
-          startFieldRightSide: icon,
-        })}
         renderAdditionalControls={withAdditionalControls ? additionalControls : undefined}
       />
     </div>
