+++ src/hooks/useForkRef/useForkRef.tsx
@@ -2,10 +2,7 @@ import { useMemo } from 'react';
 
 import { setRef } from '../../utils/setRef';
 
-type AtributesRef<T> = React.Ref<T> | undefined;
-type ResultRef<T> = React.RefCallback<T> | null;
-
-export function useForkRef<T>(refs: AtributesRef<T>[]): ResultRef<T> {
+export function useForkRef<T>(refs: (React.Ref<T> | undefined)[]): React.RefCallback<T> | null {
   return useMemo(() => {
     if (refs.length < 1) {
       return null;
